feat: bye bye Tools too!
