+++ Sources/Breeze/Repository/User/User.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Breeze\Repository\User;
 
 class User
@@ -42,8 +44,8 @@ class User
 		// Has it been defined yet?
 		if (!isset($_SESSION['Breeze_floodControl' . $user]))
 			$_SESSION['Breeze_floodControl' . $user] = [
-				'time' => time() + $seconds,
-				'msg' => 0,
+			    'time' => time() + $seconds,
+			    'msg' => 0,
 			];
 
 		// Keep track of it.
@@ -93,10 +95,10 @@ class User
 		{
 			// Set an empty array.
 			static::$_users[$u] = [
-				'breezeFacebox' => '',
-				'link' => '',
-				'name' => '',
-				'linkFacebox' => '',
+			    'breezeFacebox' => '',
+			    'link' => '',
+			    'name' => '',
+			    'linkFacebox' => '',
 			];
 
 			// Gotta make sure you're only loading info from real existing members...
@@ -119,10 +121,10 @@ class User
 			// Not a real member, fill out some guest generic vars and be done with it..
 			else
 				static::$_users[$u] = [
-					'breezeFacebox' => $txt['guest_title'],
-					'link' => $txt['guest_title'],
-					'name' => $txt['guest_title'],
-					'linkFacebox' => $txt['guest_title'],
+				    'breezeFacebox' => $txt['guest_title'],
+				    'link' => $txt['guest_title'],
+				    'name' => $txt['guest_title'],
+				    'linkFacebox' => $txt['guest_title'],
 				];
 		}
 
@@ -132,5 +134,4 @@ class User
 		if ($returnID)
 			return $loadedIDs;
 	}
-
 }
