+++ Sources/Breeze/Services/Base.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Breeze\Service;
+
+class Base
+{
+	public function global(string $variableName)
+	{
+		return $GLOBALS[$variableName] ?? false;
+	}
+}
