+++ Sources/Breeze/Services/Text.php
@@ -1,8 +1,9 @@
 <?php
 
+declare(strict_types=1);
+
 
 namespace Breeze\Service;
-
 
 class Text
 {
@@ -16,7 +17,7 @@ class Text
 		$this->tools = $tools;
 	}
 
-	public function setLanguage()
+	public function setLanguage(): void
 	{
 	}
 
@@ -25,6 +26,99 @@ class Text
 		$txt = $this->tools->global('txt');
 
 		return $txt[$textKey];
+	}
+
+	public function parser(string $text, array $replacements = []): string
+	{
+		$context = $this->global('context');
+
+		if (empty($text) || empty($replacements) || !is_array($replacements))
+			return '';
+
+		$s = ';' . $context['session_var'] . '=' . $context['session_id'];
+
+		$find = [];
+		$replace = [];
+
+		foreach ($replacements as $f => $r)
+		{
+			$find[] = '{' . $f . '}';
+			$replace[] = $r . ((false !== strpos($f, 'href')) ? $s : '');
+		}
+
+		return str_replace($find, $replace, $text);
+	}
+
+	public function commaSeparated(string $string, string $type = 'alphanumeric'): string
+	{
+		switch ($type) {
+			case 'numeric':
+				$t = '\d';
+				break;
+			case 'alpha':
+				$t = '[:alpha:]';
+				break;
+			case 'alphanumeric':
+			default:
+				$t = '[:alnum:]';
+				break;
+		}
+
+		return empty($string) ? '' : implode(',', array_filter(explode(',', preg_replace(
+		    [
+		        '/[^' . $t . ',]/',
+		        '/(?<=,),+/',
+		        '/^,+/',
+		        '/,+$/'
+		    ],
+		    '',
+		    $string
+		))));
+	}
+
+	public function normalizeString(string $string = ''): string
+	{
+		$smcFunc = $this->global('smcFunc');
+
+		if (empty($string))
+			return '';
+
+		$string = $smcFunc['htmlspecialchars']($string, \ENT_QUOTES);
+		$string = preg_replace('~&([a-z]{1,2})(acute|cedil|circ|grave|lig|orn|ring|slash|th|tilde|uml);~i', '$1', $string);
+		$string = html_entity_decode($string, \ENT_QUOTES);
+		$string = preg_replace(['~[^0-9a-z]~i', '~[ -]+~'], ' ', $string);
+
+		return trim($string, ' -');
+	}
+
+	public function formatBytes(int $bytes, bool $showUnits = false): string
+	{
+		$units = ['B', 'KB', 'MB', 'GB', 'TB'];
+
+		$bytes = max($bytes, 0);
+		$pow = floor(($bytes ? log($bytes) : 0) / log(1024));
+		$pow = min($pow, count($units) - 1);
+		$bytes /= (1 << (10 * $pow));
+
+		return round($bytes, 4) . ($showUnits ? ' ' . $units[$pow] : '');
+	}
+
+	public function truncate(string $string, int $limit, string $break = ' ', string $pad = '...'): string
+	{
+		if(empty($string))
+			return '';
+
+		if(empty($limit))
+			$limit = 30;
+
+		if(strlen($string) <= $limit)
+			return $string;
+
+		if(false !== ($breakpoint = strpos($string, $break, $limit)))
+			if($breakpoint < strlen($string) - 1)
+				$string = substr($string, 0, $breakpoint) . $pad;
+
+		return $string;
 	}
 
 	public function timeElapsed($ptime): string
@@ -36,12 +130,12 @@ class Text
 			return $txt['time_just_now'];
 
 		$a = [
-			12 * 30 * 24 * 60 * 60	=> $txt['time_year'],
-			30 * 24 * 60 * 60		=> $txt['time_month'],
-			24 * 60 * 60			=> $txt['time_day'],
-			60 * 60					=> $txt['time_hour'],
-			60						=> $txt['time_minute'],
-			1						=> $txt['time_second']
+		    12 * 30 * 24 * 60 * 60	=> $txt['time_year'],
+		    30 * 24 * 60 * 60		=> $txt['time_month'],
+		    24 * 60 * 60			=> $txt['time_day'],
+		    60 * 60					=> $txt['time_hour'],
+		    60						=> $txt['time_minute'],
+		    1						=> $txt['time_second']
 		];
 
 		foreach ($a as $secs => $str)
@@ -55,5 +149,4 @@ class Text
 			}
 		}
 	}
-
 }
